Fix build after moving resume files
