+++ src/components/Layout/index.tsx
@@ -3,10 +3,10 @@ import { graphql, useStaticQuery } from 'gatsby';
 import { ThemeProvider } from 'styled-components';
 import { CssBaseline } from '@mui/material';
 
-import { HeadMetaContext, PageMapContext } from '../../common/contexts';
-import { Footer, Header, LeftSideMenu } from '../../components';
-import { GlobalStyles, pageMap } from '../../constants';
-import { baseTheme, containerTheme, resumeTheme } from '../../themes';
+import { HeadMetaContext, PageMapContext } from 'common/contexts';
+import { Footer, Header, LeftSideMenu } from 'components/index';
+import { GlobalStyles, pageMap } from 'constants/index';
+import { baseTheme, containerTheme, resumeTheme } from 'themes/index';
 import { StyledContainerOnly, StyledBoxOnly, StyledContainer, StyledBox } from './styled';
 
 interface LayoutProps {
