--- src/resume/index.tsx
@@ -1,4 +1,0 @@
-import ContactInfo from './ContactInfo';
-import HeadingDetails from './HeadingDetails';
-
-export { ContactInfo, HeadingDetails };
