+++ tsconfig.json
@@ -19,7 +19,7 @@
             "esnext"
         ],
         "typeRoots": [".node_modules/@types", "./types/*"],
-        "types": ["@types/jest"],
+        // "types": ["@types/jest"],
         // "jsx": "react-jsx" /* Specify what JSX code is generated. */,
         "jsx": "preserve",
         // "experimentalDecorators": true,                   /* Enable experimental support for TC39 stage 2 draft decorators. */
@@ -37,7 +37,16 @@
         // "rootDir": "./",                                  /* Specify the root folder within your source files. */
         "moduleResolution": "node" /* Specify how TypeScript looks up a file from a given module specifier. */,
         // "baseUrl": "./",                                  /* Specify the base directory to resolve non-relative module names. */
-        // "paths": {},                                      /* Specify a set of entries that re-map imports to additional lookup locations. */
+        "paths": {
+            /* Specify a set of entries that re-map imports to additional lookup locations. */
+            "common/*": ["./src/common/*"],
+            "components/*": ["./src/components/*"],
+            "constants/*": ["./src/constants/*"],
+            "pages/*": ["./src/pages/*"],
+            "resume/*": ["./src/resume/*"],
+            "themes/*": ["./src/themes/*"],
+            "utils/*": ["./src/utils/*"]
+        },
         // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
         // "typeRoots": [],                                  /* Specify multiple folders that act like `./node_modules/@types`. */
         // "types": [],                                      /* Specify type package names to be included without being referenced in a source file. */
